+++ src/lib/getLogger.js
@@ -3,6 +3,7 @@ const path = require('path');
 let logger = {
   log: console.log.bind(console),
 };
+
 try {
   const file = path.join(process.cwd(), '/logger.js');
   console.log('Loading ', file);