+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 An overview of the changes made in various versions.
 
+### 4.0.0
+   - String based scrapers
+   - scrape() now accepts a queueItem
+   - removed setParams
+   - removed filterParams
+
 ### 3.0.1
   - Fix logs